+++ Code/infochange/client/src/pages/dashboard/Dashboard.jsx
@@ -7,11 +7,13 @@ import {
     Wallet2,
     LayoutTextSidebar,
     Send,
+    Gear
 } from "react-bootstrap-icons";
 import Profile from "./windows/Profile";
 import Wallet from "./windows/Wallet";
 import History from "./windows/history/History";
 import Bizum from "./windows/bizum/Bizum";
+import Configuration from "./windows/Configuration";
 
 import { useAuth } from "../authenticator/AuthContext";
 import { useAPI } from "../../context/APIContext";
@@ -56,7 +58,7 @@ function Dashboard() {
     }
 
     let pages = [
-        <Profile profile={user.profile} swap={doSwap} />,
+        <Profile profile={user.profile} />,
         <Wallet wallet={user.wallet ?? {}} />,
         <History
             tradeHistory={tradeHistory}
@@ -65,9 +67,10 @@ function Dashboard() {
             user={user}
         />,
         <Bizum user={user} bizumUsers={bizumUsers} reload={loadBizumHistory} />,
+        <Configuration profile={user.profile} swap={doSwap} />
     ];
 
-    const labels = ["Perfil", "Cartera", "Historial", "Bizum"];
+    const labels = ["Perfil", "Cartera", "Historial", "Bizum", "Configuración"];
 
     return (
         <div className="container">
@@ -79,18 +82,16 @@ function Dashboard() {
                     <div className="list-group">
                         <button
                             type="button"
-                            className={`list-group-item list-group-item-action ${
-                                page === 0 ? "active" : ""
-                            } d-flex align-items-center`}
+                            className={`list-group-item list-group-item-action ${page === 0 ? "active" : ""
+                                } d-flex align-items-center`}
                             onClick={() => setPage(0)}
                         >
                             <PersonFill className="me-3" /> Perfil
                         </button>
                         <button
                             type="button"
-                            className={`list-group-item list-group-item-action ${
-                                page === 1 ? "active" : ""
-                            } d-flex align-items-center`}
+                            className={`list-group-item list-group-item-action ${page === 1 ? "active" : ""
+                                } d-flex align-items-center`}
                             onClick={() => setPage(1)}
                         >
                             <Wallet2 className="me-3" />
@@ -98,9 +99,8 @@ function Dashboard() {
                         </button>
                         <button
                             type="button"
-                            className={`list-group-item list-group-item-action ${
-                                page === 2 ? "active" : ""
-                            } d-flex align-items-center`}
+                            className={`list-group-item list-group-item-action ${page === 2 ? "active" : ""
+                                } d-flex align-items-center`}
                             onClick={() => setPage(2)}
                         >
                             <LayoutTextSidebar className="me-3" />
@@ -108,13 +108,21 @@ function Dashboard() {
                         </button>
                         <button
                             type="button"
-                            className={`list-group-item list-group-item-action ${
-                                page === 3 ? "active" : ""
-                            } d-flex align-items-center`}
+                            className={`list-group-item list-group-item-action ${page === 3 ? "active" : ""
+                                } d-flex align-items-center`}
                             onClick={() => setPage(3)}
                         >
                             <Send className="me-3" />
                             Bizum
+                        </button>
+                        <button
+                            type="button"
+                            className={`list-group-item list-group-item-action ${page === 4 ? "active" : ""
+                                } d-flex align-items-center`}
+                            onClick={() => setPage(4)}
+                        >
+                            <Gear className="me-3" />
+                            Ajustes
                         </button>
                     </div>
                 </aside>
