+++ Code/infochange/client/src/pages/dashboard/windows/Profile.jsx
@@ -1,5 +1,5 @@
 export default function Profile(props) {
-    const { profile, swap } = props;
+    const { profile } = props;
 
     const personalInfo = [
         {
@@ -43,14 +43,6 @@ export default function Profile(props) {
         },
     ];
 
-    const handleSwap = async () => {
-        const loadingScreen = document.getElementById("loading-screen");
-
-        loadingScreen.style.display = "block";
-        await swap();
-        loadingScreen.style.display = "none";
-    };
-
     return (
         <>
             <div className="row px-5 py-4">
@@ -65,8 +57,8 @@ export default function Profile(props) {
                         className="rounded-pill"
                         alt={`Foto de perfil de ${profile.username}`}
                         onError={(event) =>
-                            (event.target.src =
-                                "https://static.vecteezy.com/system/resources/thumbnails/009/292/244/small/default-avatar-icon-of-social-media-user-vector.jpg")
+                        (event.target.src =
+                            "https://static.vecteezy.com/system/resources/thumbnails/009/292/244/small/default-avatar-icon-of-social-media-user-vector.jpg")
                         }
                         src={`https://github.com/${profile.username}.png`}
                         style={{ width: "100%" }}
@@ -88,26 +80,6 @@ export default function Profile(props) {
                         <h3 className="fs-4 text-body-secondary">
                             @{profile.username}
                         </h3>
-                    </div>
-                    <div className="col-6 d-flex justify-content-start align-items-end flex-column">
-                        <div>
-                            <button
-                                className="btn btn-primary mt-2 mb-2 d-flex"
-                                onClick={() => handleSwap()}
-                            >
-                                Cambiar a modo{" "}
-                                {profile.mode === 0 ? "profesional" : "novato"}
-                            </button>
-                        </div>
-                        <span className="text-end">
-                            {profile.mode === 0
-                                ? "Con el modo profesional, el gráfico de trading tendrá muchas más opciones, " +
-                                  "además de que mostrará una mayor cantidad de pares. " +
-                                  "También se muestran más opciones a la hora de introducir la cantidad de compra o venta deseada " +
-                                  "No se recomienda para usuarios inexpertos."
-                                : "Con el modo novato, la cantidad de pares de trading están limitados a los que tienen relación con " +
-                                  "el dólar estadounidense. Además, el gráfico se simplifica y la forma de comprar y vender es más básica."}
-                        </span>
                     </div>
                 </div>
             </div>
